pyproject.toml
[tool.poetry]
name = "talk-trek-server"
version = "0.1.0"
description = ""
authors = ["Bill1907 <a31829660@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
requests = "^2.31.0"
uvicorn = "^0.23.2"
aiofiles = "^23.1.0"
websockets = "^12.0"
openai = "^1.33.0"
python-dotenv = "^1.0.1"
supabase = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"